Reactor on Instruction Cycles (#20)
Changed the way in which MCT cycles are being reported for each executed instruction. Prior to this, each instruction would modify a public structure variable, cycles. Since the return value for step and associated instruction execution (bool) was not being used. It was refactored to return the number of cycles executed u16.